Replace dataclasses
